@@ -18,15 +18,23 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
+  cloud_firestore: ^2.2.0
   cupertino_icons: ^1.0.2
-  firebase_core: ^0.7.0
   dio: ^4.0.0
-
+  firebase_auth: ^1.2.0
+  firebase_core: ^1.2.0
+  flutter_bloc: ^7.0.0
+  get_it: ^7.1.3
+  hydrated_bloc: ^7.0.0
+  modal_progress_hud_nsn: ^0.1.0-nullsafety-1
+  path_provider: ^2.0.2
+  rflutter_alert: ^2.0.2
+  shared_preferences: ^2.0.6
 
 dev_dependencies:
   flutter_test:
