@@ -1,66 +1,54 @@
 class FilmlerCevap {
-  int page;
-  List<Film> filmlerListesi;
-  int totalPages;
-  int totalResults;
+  int? page;
+  List<Film>? filmlerListesi;
+  int? totalPages;
+  int? totalResults;
 
-  FilmlerCevap({this.page, this.filmlerListesi, this.totalPages, this.totalResults});
+  FilmlerCevap(
+      {this.page, this.filmlerListesi, this.totalPages, this.totalResults});
 
   FilmlerCevap.fromJson(Map<String, dynamic> json) {
     page = json['page'];
     if (json['results'] != null) {
       filmlerListesi = <Film>[];
       json['results'].forEach((v) {
-        filmlerListesi.add(new Film.fromJson(v));
+        filmlerListesi?.add(new Film.fromJson(v));
       });
     }
     totalPages = json['total_pages'];
     totalResults = json['total_results'];
   }
-
-  Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = new Map<String, dynamic>();
-    data['page'] = this.page;
-    if (this.filmlerListesi != null) {
-      data['results'] = this.filmlerListesi.map((v) => v.toJson()).toList();
-    }
-    data['total_pages'] = this.totalPages;
-    data['total_results'] = this.totalResults;
-    return data;
-  }
 }
 
 class Film {
-  bool adult;
-  String backdropPath;
-  Null belongsToCollection;
-  int budget;
-  List<Genres> genres;
-  String homepage;
-  int id;
-  String imdbId;
-  String originalLanguage;
-  String originalTitle;
-  String overview;
-  double popularity;
-  String posterPath;
-  List<ProductionCompanies> productionCompanies;
-  List<ProductionCountries> productionCountries;
-  String releaseDate;
-  int revenue;
-  int runtime;
-  List<SpokenLanguages> spokenLanguages;
-  String status;
-  String tagline;
-  String title;
-  bool video;
-  double voteAverage;
-  int voteCount;
+  bool? adult;
+  String? backdropPath;
+  int? budget;
+  List<Genres>? genres;
+  String? homepage;
+  int? id;
+  String? imdbId;
+  String? originalLanguage;
+  String? originalTitle;
+  String? overview;
+  double? popularity;
+  String? posterPath;
+  List<ProductionCompanies>? productionCompanies;
+  List<ProductionCountries>? productionCountries;
+  String? releaseDate;
+  int? revenue;
+  int? runtime;
+  List<SpokenLanguages>? spokenLanguages;
+  String? status;
+  String? tagline;
+  String? title;
+  bool? video;
+  double? voteAverage;
+  int? voteCount;
 
   Film(
       {this.adult,
       this.backdropPath,
-      this.belongsToCollection,
       this.budget,
       this.genres,
       this.homepage,
@@ -87,12 +75,11 @@ class Film {
   Film.fromJson(Map<String, dynamic> json) {
     adult = json['adult'];
     backdropPath = json['backdrop_path'];
-    belongsToCollection = json['belongs_to_collection'];
     budget = json['budget'];
     if (json['genres'] != null) {
       genres = <Genres>[];
       json['genres'].forEach((v) {
-        genres.add(new Genres.fromJson(v));
+        genres?.add(new Genres.fromJson(v));
       });
     }
     homepage = json['homepage'];
@@ -106,13 +93,13 @@ class Film {
     if (json['production_companies'] != null) {
       productionCompanies = <ProductionCompanies>[];
       json['production_companies'].forEach((v) {
-        productionCompanies.add(new ProductionCompanies.fromJson(v));
+        productionCompanies?.add(new ProductionCompanies.fromJson(v));
       });
     }
     if (json['production_countries'] != null) {
       productionCountries = <ProductionCountries>[];
       json['production_countries'].forEach((v) {
-        productionCountries.add(new ProductionCountries.fromJson(v));
+        productionCountries?.add(new ProductionCountries.fromJson(v));
       });
     }
     releaseDate = json['release_date'];
@@ -121,7 +108,7 @@ class Film {
     if (json['spoken_languages'] != null) {
       spokenLanguages = <SpokenLanguages>[];
       json['spoken_languages'].forEach((v) {
-        spokenLanguages.add(new SpokenLanguages.fromJson(v));
+        spokenLanguages?.add(new SpokenLanguages.fromJson(v));
       });
     }
     status = json['status'];
@@ -131,52 +118,11 @@ class Film {
     voteAverage = json['vote_average'];
     voteCount = json['vote_count'];
   }
-
-  Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = new Map<String, dynamic>();
-    data['adult'] = this.adult;
-    data['backdrop_path'] = this.backdropPath;
-    data['belongs_to_collection'] = this.belongsToCollection;
-    data['budget'] = this.budget;
-    if (this.genres != null) {
-      data['genres'] = this.genres.map((v) => v.toJson()).toList();
-    }
-    data['homepage'] = this.homepage;
-    data['id'] = this.id;
-    data['imdb_id'] = this.imdbId;
-    data['original_language'] = this.originalLanguage;
-    data['original_title'] = this.originalTitle;
-    data['overview'] = this.overview;
-    data['popularity'] = this.popularity;
-    data['poster_path'] = this.posterPath;
-    if (this.productionCompanies != null) {
-      data['production_companies'] =
-          this.productionCompanies.map((v) => v.toJson()).toList();
-    }
-    if (this.productionCountries != null) {
-      data['production_countries'] =
-          this.productionCountries.map((v) => v.toJson()).toList();
-    }
-    data['release_date'] = this.releaseDate;
-    data['revenue'] = this.revenue;
-    data['runtime'] = this.runtime;
-    if (this.spokenLanguages != null) {
-      data['spoken_languages'] =
-          this.spokenLanguages.map((v) => v.toJson()).toList();
-    }
-    data['status'] = this.status;
-    data['tagline'] = this.tagline;
-    data['title'] = this.title;
-    data['video'] = this.video;
-    data['vote_average'] = this.voteAverage;
-    data['vote_count'] = this.voteCount;
-    return data;
-  }
 }
 
 class Genres {
-  int id;
-  String name;
+  int? id;
+  String? name;
 
   Genres({this.id, this.name});
 
@@ -194,10 +140,10 @@ class Genres {
 }
 
 class ProductionCompanies {
-  int id;
-  String logoPath;
-  String name;
-  String originCountry;
+  int? id;
+  String? logoPath;
+  String? name;
+  String? originCountry;
 
   ProductionCompanies({this.id, this.logoPath, this.name, this.originCountry});
 
@@ -219,8 +165,8 @@ class ProductionCompanies {
 }
 
 class ProductionCountries {
-  String iso31661;
-  String name;
+  String? iso31661;
+  String? name;
 
   ProductionCountries({this.iso31661, this.name});
 
@@ -238,9 +184,9 @@ class ProductionCountries {
 }
 
 class SpokenLanguages {
-  String englishName;
-  String iso6391;
-  String name;
+  String? englishName;
+  String? iso6391;
+  String? name;
 
   SpokenLanguages({this.englishName, this.iso6391, this.name});
 
