@@ -1,6 +1,15 @@
-const String moviesUrl ='https://api.themoviedb.org/3/discover/movie?api_key=ba99a0214490746fac583d2656a662e2&page=1';
+import 'package:flutter/material.dart';
+
+const String moviesUrl =
+    'https://api.themoviedb.org/3/discover/movie?api_key=ba99a0214490746fac583d2656a662e2&page=1';
 
 const String baseImageUrl = 'https://image.tmdb.org/t/p/original/';
 
-const String filmUrl = 'https://api.themoviedb.org/3/movie/632357?api_key=ba99a0214490746fac583d2656a662e2';
+const String filmUrl =
+    'https://api.themoviedb.org/3/movie/632357?api_key=ba99a0214490746fac583d2656a662e2';
 
+const kSkinColor = Color(0xffF4F1F1);
+
+const kBlueColor = Color(0xff00B4E4);
+
+const String kOnboarding = 'onboarding';
