@@ -1,8 +1,6 @@
-
 import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
 import 'package:my_app/core/const.dart';
-import 'package:my_app/models/film.dart';
 import 'package:my_app/models/movies.dart';
 import 'package:my_app/screens/detail_screen.dart';
 
@@ -32,9 +30,9 @@ class _HomeScreenState extends State<HomeScreen> {
   Future<MovieResponse> filmlerGetir() async {
     var movieResponse;
     try {
-       var response = await Dio().get(moviesUrl);
-       movieResponse = MovieResponse.fromJson(response.data);
-       print(response);
+      var response = await Dio().get(moviesUrl);
+      movieResponse = MovieResponse.fromJson(response.data);
+      print(response);
     } catch (e) {
       print(e);
     }
@@ -58,37 +56,42 @@ class _HomeScreenState extends State<HomeScreen> {
             var movieResponse = snapshot.data;
 
             return GridView.builder(
-              itemCount: movieResponse.movies.length,
+              itemCount: movieResponse!.movies!.length,
               gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
                 crossAxisCount: 2,
                 childAspectRatio: 2 / 3.5,
               ),
               itemBuilder: (context, indeks) {
-                var movie = movieResponse.movies[indeks];
+                var movie = movieResponse.movies![indeks];
 
                 return GestureDetector(
-                  onTap: (){
-                    Navigator.push(context, MaterialPageRoute(builder: (context)=>DetailScreen(film: film,)));
+                  onTap: () {
+                    Navigator.push(
+                        context,
+                        MaterialPageRoute(
+                            builder: (context) => DetailScreen(
+                                  movie: movie!,
+                                )));
                   },
-                                  child: Card(
+                  child: Card(
                     child: Column(
                       mainAxisAlignment: MainAxisAlignment.spaceEvenly,
                       children: [
                         Expanded(
                             child: Padding(
-                              padding: const EdgeInsets.all(8.0),
-                              child: Image.network(
-                          '$baseImageUrl/${movie.backdropPath}',
-                          fit: BoxFit.cover,
-                        ),
-                            )),
+                          padding: const EdgeInsets.all(8.0),
+                          child: Image.network(
+                            '$baseImageUrl/${movie!.backdropPath}',
+                            fit: BoxFit.cover,
+                          ),
+                        )),
                         Text(
-                          movie.title,
+                          movie.title!,
                           style: TextStyle(
                               fontSize: 16, fontWeight: FontWeight.bold),
                         ),
                         Text(
-                          movie.releaseDate,
+                          movie.releaseDate!,
                           style: TextStyle(
                             fontSize: 14,
                           ),
